Add setting for extended attributes filter and start using changes since token

+++ handler/src/main/java/org/n52/sos/ds/observation/ObservationValueCreator.java
@@ -68,7 +68,7 @@ public class ObservationValueCreator extends AbstractObservationValueCreator {
 
     @Override
     public QuantityValue visit(QuantityDataEntity o) {
-        return visit(o, new QuantityValue(o.hasValue() ? o.getValue().doubleValue() : null));
+        return visit(o, new QuantityValue(o.hasValue() ? o.getValue() : null));
     }
 
     @Override
